fix(#234): get_inventory can never verify an empty chain it cannot see
Zero enumerated slots now returns State B readback_unavailable with
plugins_unknown_reason=insert_section_not_enumerable instead of a
false verified-empty State A (a visible insert section always exposes
at least the empty append row). Write-path slot guards share one
zero-slot detail so "(0 slots)" failures name the real condition.
